+++ src/main/java/cn/nukkit/block/BlockLeaves.java
@@ -7,8 +7,10 @@ import cn.nukkit.item.Item;
 import cn.nukkit.item.ItemTool;
 import cn.nukkit.level.Level;
 import cn.nukkit.math.BlockFace;
-import cn.nukkit.math.SimpleAxisAlignedBB;
 import cn.nukkit.utils.BlockColor;
+import cn.nukkit.utils.Hash;
+import it.unimi.dsi.fastutil.longs.Long2LongMap;
+import it.unimi.dsi.fastutil.longs.Long2LongOpenHashMap;
 
 import java.util.concurrent.ThreadLocalRandom;
 
@@ -17,6 +19,9 @@ import java.util.concurrent.ThreadLocalRandom;
  * Nukkit Project
  */
 public class BlockLeaves extends BlockTransparentMeta {
+    private static final BlockFace[] VISIT_ORDER = new BlockFace[]{
+            BlockFace.NORTH, BlockFace.SOUTH, BlockFace.EAST, BlockFace.WEST, BlockFace.DOWN, BlockFace.UP
+    };
     public static final int OAK = 0;
     public static final int SPRUCE = 1;
     public static final int BIRCH = 2;
@@ -114,7 +119,7 @@ public class BlockLeaves extends BlockTransparentMeta {
     public int onUpdate(int type) {
         if (type == Level.BLOCK_UPDATE_RANDOM) {
             if (isCheckDecay()) {
-                if (isPersistent() || findLog(this, 7)) {
+                if (isPersistent() || findLog(this, 7, null)) {
                     setCheckDecay(false);
                     getLevel().setBlock(this, this, false, false);
                 } else {
@@ -147,16 +152,28 @@ public class BlockLeaves extends BlockTransparentMeta {
         return type;
     }
 
-    private Boolean findLog(Block pos, Integer distance) {
-        Block[] woodBlocks = this.getLevel().getCollisionBlocks(
-                new SimpleAxisAlignedBB(
-                        pos.getX() - distance, pos.getY() - distance, pos.getZ() - distance,
-                        pos.getX() + distance, pos.getY() + distance, pos.getZ() + distance
-                ), true, true,
-                block -> block.getId() == WOOD || block.getId() == WOOD2
-        );
-        
-        return woodBlocks.length > 0;
+    private Boolean findLog(Block current, int distance, Long2LongMap visited) {
+        if (visited == null) {
+            visited = new Long2LongOpenHashMap();
+            visited.defaultReturnValue(-1);
+        }
+        if (current instanceof BlockWood) {
+            return true;
+        }
+        if (distance == 0 || !(current instanceof BlockLeaves)) {
+            return false;
+        }
+        long hash = Hash.hashBlock(current);
+        if (visited.get(hash) >= distance) {
+            return false;
+        }
+        visited.put(hash, distance);
+        for (BlockFace face : VISIT_ORDER) {
+            if(findLog(current.getSide(face), distance - 1, visited)) {
+                return true;
+            }
+        }
+        return false;
     }
 
     public boolean isCheckDecay() {