+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Click the link above to see the future.
 - [#244] `Enchantment.getPowerNukkit().isItemAcceptable(Item)` to check if an enchantment can exist 
          in a given item stack by any non-hack means.
 - [#256] `CapturingCommandSender` intended to capture output of commands which don't require players.
+- [#259] `Hash.hashBlock(Vector3)` method for public usage.
 
 ### Changed
 - [#227] Sugar canes now fires BlockGrowEvent when growing naturally.
@@ -38,7 +39,9 @@ Click the link above to see the future.
 - [#193] Wrong explosion behaviour with waterlogged block
 - [#103] Fixes BlockLeaves's random update logic spamming packets and consuming CPU unnecessarily
 - [#253] Fixes `LeavesDecayEvent` also being called when leaves wouldn't decay
+- [#254] Fixes BlockLeaves not checking for log connectivity, was checking only if it had a log block nearby
 - [#255] Fix /status information in /debugpaste not being collected
+
 
 ### Changed
 - [#247] Invalid BlockId:Meta combinations now log an error when found. It logs only once
@@ -260,5 +263,7 @@ Fixes several anvil issues.
 [#247]: https://github.com/GameModsBR/PowerNukkit/pull/247
 [#248]: https://github.com/GameModsBR/PowerNukkit/pull/248
 [#253]: https://github.com/GameModsBR/PowerNukkit/pull/253
+[#254]: https://github.com/GameModsBR/PowerNukkit/issues/254
 [#255]: https://github.com/GameModsBR/PowerNukkit/pull/255
 [#256]: https://github.com/GameModsBR/PowerNukkit/pull/256
+[#259]: https://github.com/GameModsBR/PowerNukkit/pull/259